+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Internal changes] 4.10.2021 - 1
+* Texture Cleanup
+
 ## [Fixes] 4.5.2021 - 2
 * Fixed crash during startup